requirements.txt
requests>=2.31.0
pandas==2.2.2
Pillow==10.3.0
numpy==1.26.4
scikit-learn>=1.7.2
tensorflow==2.16.1
matplotlib==3.8.4
mlflow>=2.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
boto3>=1.29.0